+++ packages/react/src/components/Accordion/components/AccordionProvider.tsx
@@ -1,9 +1,9 @@
-import React, { ReactNode } from 'react';
-import { AccordionContext, AccordionProviderValue } from '../utils/constants';
+import * as React from "react";
+import { AccordionContext, AccordionProviderValue } from "../utils/constants";
 
 const AccordionProvider = (props: {
   value: AccordionProviderValue;
-  children?: ReactNode;
+  children?: React.ReactNode;
 }) => <AccordionContext.Provider {...props} />;
 
 export { AccordionProvider };