+++ packages/react/src/components/StepNavigation/StepNavigation.tsx
@@ -40,7 +40,6 @@ interface StepNavigationProps {
 }
 
 function StepNavigation({
-  children,
   inline,
   small,
   vertical,