+++ packages/website/components/Blog/References/index.tsx
@@ -1,9 +1,9 @@
-import { faFigma, faGithub, faNpm } from '@fortawesome/free-brands-svg-icons';
-import { faCode } from '@fortawesome/free-solid-svg-icons';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { Link, List, ListItem } from '@wfp/react';
-import React from 'react';
-import styles from './tableOfContent.module.scss';
+import { faFigma, faGithub, faNpm } from "@fortawesome/free-brands-svg-icons";
+import { faCode } from "@fortawesome/free-solid-svg-icons";
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+import { Link, List, ListItem } from "@wfp/react";
+import React from "react";
+import styles from "./tableOfContent.module.scss";
 
 export default function References({ post }: any) {
   if (!post.figma && !post.storybook && !post.github) return null;
@@ -18,24 +18,34 @@ export default function References({ post }: any) {
             </Link>
           </ListItem>
         )}
-        {post.github && (
+
+        {post.slug.split("/").includes("Components") && (
           <ListItem>
-            <Link href={post.github} target="_blank">
-              GitHub{' '}
+            <Link
+              href={`${
+                process.env.NEXT_PUBLIC_GIT_URL
+              }/packages/react/src/components/${post.title.replace(
+                /[/]/g,
+                "-"
+              )}`}
+              target="_blank"
+            >
+              Source{" "}
               <FontAwesomeIcon icon={faGithub} className={styles.github} />
             </Link>
           </ListItem>
         )}
-        {post.storybook && (
+        {post.slug.split("/").includes("Components") && (
           <ListItem>
             <Link
-              href={`http://localhost:9001/?path=/story/${
+              href={`${process.env.NEXT_PUBLIC_STORYBOOK_URL}/?path=/story/${
                 post.storybook == true
-                  ? post.slug.replace(/[/]/g, '-')
+                  ? post.slug.replace(/[/]/g, "-")
                   : post.storybook
               }`}
-              target="_blank">
-              Storybook{' '}
+              target="_blank"
+            >
+              Storybook{" "}
               <FontAwesomeIcon icon={faCode} className={styles.storybook} />
             </Link>
           </ListItem>
@@ -45,7 +55,8 @@ export default function References({ post }: any) {
           <ListItem>
             <Link
               href={`https://www.npmjs.com/package/${post.npm}`}
-              target="_blank">
+              target="_blank"
+            >
               npm <FontAwesomeIcon icon={faNpm} className={styles.npm} />
             </Link>
           </ListItem>