+++ packages/website/components/Post/getPostContent.tsx
@@ -30,6 +30,7 @@ export default async function getPostContent(params: any) {
     "previewScale",
     "mainComponent",
     "components",
+    "componentsNew",
     "defaultProps",
     "sampleCode",
   ]);
@@ -59,6 +60,7 @@ export default async function getPostContent(params: any) {
         "coverImage",
         "mainComponent",
         "components",
+        "componentsNew",
         "defaultProps",
         "sampleCode",
         "excerpt",
@@ -92,6 +94,22 @@ export default async function getPostContent(params: any) {
     } catch (e) {
       console.log("Can't load typescript definitions!");
     }
+  }
+
+  if (post.componentsNew) {
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
+    Object.entries(
+      post.componentsNew as { [key: string]: { path?: string } }
+      // eslint-disable-next-line @typescript-eslint/no-unused-vars
+    ).map(([key, component]) => {
+      try {
+        // eslint-disable-next-line @typescript-eslint/no-var-requires
+        const file = require(`../../types/src/components/${component.path}.json`);
+        propTypes.push(file[0]);
+      } catch (e) {
+        console.log("Can't load typescript definitions!");
+      }
+    });
   }
 
   if (post.components) {