+++ packages/website/components/PropTypes/index.tsx
@@ -59,6 +59,7 @@ export default function PropTypes({
   components = [],
   //sampleCode: sampleCodeInput,
   //smallPreview,
+  showEditor,
   previewScale,
   propTypes,
   view,
@@ -258,6 +259,7 @@ export default function PropTypes({
           // smallPreview
           live
           view={view}
+          showEditor={showEditor}
           // showEditor={!showAllProps}
         />
       </div>