+++ packages/icons/CHANGELOG.md
@@ -3,6 +3,60 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 10.35.7 (2023-12-01)
+
+**Note:** Version bump only for package @un/icons
+
+
+
+
+
+## 10.35.6 (2023-12-01)
+
+**Note:** Version bump only for package @un/icons
+
+
+
+
+
+## 10.35.5 (2023-12-01)
+
+**Note:** Version bump only for package @un/icons
+
+
+
+
+
+## 10.35.4 (2023-11-30)
+
+
+### Bug Fixes
+
+* **packages:** fixed yarn lock errors and inconsitencies in themes ([14ace8c](https://github.com/wfp/designsystem/commit/14ace8cc950c85a8b7220c516d7eaca2cbc8df48))
+
+
+
+
+
+## 10.35.3 (2023-11-30)
+
+**Note:** Version bump only for package @un/icons
+
+
+
+
+
+## 10.35.2 (2023-11-29)
+
+
+### Bug Fixes
+
+* **Storybook:** fixed TypeScript errors and publishing it ([ca94293](https://github.com/carbon-design-system/carbon/commit/ca942938534e06d98a5799340d21aa0a58cb6847))
+
+
+
+
+
 ## 10.35.1 (2023-11-28)
 
 **Note:** Version bump only for package @un/icons